[fix] package name
