+++ pkg/payment/put_payment_payment_id.go
@@ -7,7 +7,7 @@ import (
 	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/api"
 	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/api/response"
 	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/db"
-	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/utils"
+	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/util"
 	"github.com/labstack/echo/v4"
 )
 
@@ -17,7 +17,7 @@ func PutPaymentPaymentId(ctx echo.Context, dbClient db.TransactionClient, paymen
 	if err != nil {
 		return api.ResGetPaymentPaymentId{}, err
 	}
-	err = utils.RenewalActiveLimit(dbClient, userId, strconv.Itoa(utils.GetSchoolYear()+1)+"-05-01")
+	err = util.RenewalActiveLimit(dbClient, userId, strconv.Itoa(util.GetSchoolYear()+1)+"-05-01")
 	if err != nil {
 		return api.ResGetPaymentPaymentId{}, err
 	}
