+++ pkg/user/put_user_me_renewal.go
@@ -6,13 +6,13 @@ import (
 	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/api"
 	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/api/response"
 	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/db"
-	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/utils"
+	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/util"
 	"github.com/labstack/echo/v4"
 )
 
 func PutUserMeRenewal(ctx echo.Context, dbClient db.TransactionClient) (api.ResGetUserMe, *response.Error) {
 	userId := ctx.Get("user_id").(string)
-	err := utils.RenewalActiveLimit(dbClient, userId, strconv.Itoa(utils.GetYear())+"-06-01")
+	err := util.RenewalActiveLimit(dbClient, userId, strconv.Itoa(util.GetYear())+"-06-01")
 	if err != nil {
 		return api.ResGetUserMe{}, err
 	}
