+++ pkg/util/util.go
@@ -17,6 +17,17 @@ func GetSchoolYear() int {
 	return now.Year()
 }
 
+func GetYear() int {
+	now := time.Now()
+	return now.Year()
+}
+
+func GetAfterDate(year int, month int, day int) string {
+	now := time.Now()
+	now = now.AddDate(year, month, day)
+	return now.Format("2006-01-02")
+}
+
 func GetUniqueString(str []string) []string {
 	m := make(map[string]bool)
 	uniq := []string{}
@@ -28,6 +39,21 @@ func GetUniqueString(str []string) []string {
 		}
 	}
 	return uniq
+}
+
+func RenewalActiveLimit(dbClient db.TransactionClient, userId string, activeLimit string) *response.Error {
+	params := struct {
+		UserId      string `twowaysql:"userId"`
+		ActiveLimit string `twowaysql:"activeLimit"`
+	}{
+		UserId:      userId,
+		ActiveLimit: activeLimit,
+	}
+	_, err := dbClient.Exec("sql/utils/update_user_active_limit.sql", &params, false)
+	if err != nil {
+		return &response.Error{Code: http.StatusInternalServerError, Level: "Error", Message: "不明なエラーが発生しました", Log: err.Error()}
+	}
+	return nil
 }
 
 type FileId struct {
