--- pkg/utils/utils.go
@@ -1,57 +1,0 @@
-package utils
-
-import (
-	"net/http"
-	"time"
-
-	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/api/response"
-	"github.com/SIT-DigiCre/digicore_v3_backend/pkg/db"
-)
-
-func GetSchoolYear() int {
-	now := time.Now()
-	month := int(now.Month())
-	if 1 <= month && month <= 3 {
-		return now.Year() - 1
-	}
-	return now.Year()
-}
-
-func GetYear() int {
-	now := time.Now()
-	return now.Year()
-}
-
-func GetAfterDate(year int, month int, day int) string {
-	now := time.Now()
-	now = now.AddDate(year, month, day)
-	return now.Format("2006-01-02")
-}
-
-func GetUniqueString(str []string) []string {
-	m := make(map[string]bool)
-	uniq := []string{}
-
-	for _, ele := range str {
-		if !m[ele] {
-			m[ele] = true
-			uniq = append(uniq, ele)
-		}
-	}
-	return uniq
-}
-
-func RenewalActiveLimit(dbClient db.TransactionClient, userId string, activeLimit string) *response.Error {
-	params := struct {
-		UserId      string `twowaysql:"userId"`
-		ActiveLimit string `twowaysql:"activeLimit"`
-	}{
-		UserId:      userId,
-		ActiveLimit: activeLimit,
-	}
-	_, err := dbClient.Exec("sql/utils/update_user_active_limit.sql", &params, false)
-	if err != nil {
-		return &response.Error{Code: http.StatusInternalServerError, Level: "Error", Message: "不明なエラーが発生しました", Log: err.Error()}
-	}
-	return nil
-}
